renamed sdk directory to be lowercase by pip compliance
